v1.3.2.2 RC1 Commit Change
- Time variables from Hard Cut Modes and Auto-Lock preset when it's no music are now fps-independent.
- Show the message when the preset is locked/unlocked by pressing '`' key.
- New transitions
- New waveform
- Remastered beat detection splitting algorithm (now based via sample rate) and made the configurable one via .ini file.
- Auto time variable reset and hotkey for instant reset
- time variable changed to "double"
- "Press F1 for Help" visual rework
- [BETA FIX] Transparency Mode failed on Windows 7, 8 and 8.1. It disables/reenables aero composition for better support
- More presets, textures and preset patch fixes.

+++ vis_milk2/pluginshell.h
@@ -159,7 +159,7 @@ private:
 
     // GENERAL PRIVATE STUFF
     int          m_frame;           // current frame #, starting at zero
-    float        m_time;            // current animation time in seconds; starts at zero.
+    double        m_time;            // current animation time in seconds; starts at zero.
     float        m_fps;             // current estimate of frames per second
     HINSTANCE    m_hInstance;       // handle to application instance
     DXContext*   m_lpDX;            // pointer to DXContext object
